src/server_x/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
aerich
